Add more shitty code to the already shitty code
This is for allowing Astron to run in compiled clients xd

Basically, the DC file data is currently packaged with the game code
inside of GameData.bin. In order for Astron to read the DC files, we need
it on disk. So for now, we'll just write it to a temporary directory on
every launch.